moved the Graphical UI to another package to seperate it from the logic
